# api/openapi/spec.json
{
  "openapi": "3.0.3",
...
  "info": {
    "title": "TempConnect API",
    "version": "2.0.0",
    "description": "B2B Workforce Marketplace API — VMS, Capacity Exchange, Timesheets, RBAC, Compliance.\n\nAuthentication: Session-cookie (`tc.sid`) + `X-CSRF-Token` header for mutating requests.\n\nGet a CSRF token via `GET /api/csrf` before any POST/PATCH/DELETE.\n\n⚠️ NOTICE: This spec covers only the initial pilot routes (27 of 350+ endpoints). See docs/API_DOCUMENTATION.md for current coverage.",
    "contact": {
      "name": "TempConnect Engineering",
      "email": "tech@tempconnect.de"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    { "url": "http://localhost:3000/api", "description": "Local development" },
    { "url": "https://api.tempconnect.de/api", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tc.sid",
        "description": "HTTP-only session cookie set on login"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token obtained from GET /api/csrf"
      }
    },
    "schemas": {
...
      "CapacityPost": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "supplier_company_id": { "type": "string", "format": "uuid" },
          "role": { "type": "string" },
          "skill_tags": { "type": "array", "items": { "type": "string" } },
          "headcount_available": { "type": "integer", "minimum": 1 },
          "availability_from": { "type": "string", "format": "date", "nullable": true },
          "availability_to": { "type": "string", "format": "date", "nullable": true },
          "location_city": { "type": "string", "nullable": true },
          "location_postal": { "type": "string", "nullable": true },
          "location_lat": { "type": "number", "nullable": true },
          "location_lng": { "type": "number", "nullable": true },
          "radius_km": { "type": "integer" },
          "hourly_rate_min_cents": { "type": "integer", "nullable": true },
          "hourly_rate_max_cents": { "type": "integer", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["draft", "active", "paused", "filled", "expired", "archived"] },
          "is_active": { "type": "boolean" },
          "notdienst": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [
    { "sessionCookie": [], "csrfToken": [] }
  ],
  "paths": {
...
    "/capacities": {
      "get": {
        "tags": ["capacities"],
        "summary": "List capacity posts",
        "parameters": [
          { "name": "role", "in": "query", "schema": { "type": "string" } },
          { "name": "city", "in": "query", "schema": { "type": "string" } },
          { "name": "postal_code", "in": "query", "schema": { "type": "string" } },
          { "name": "radius_km", "in": "query", "schema": { "type": "integer" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": {
            "description": "Capacity posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/CapacityPost" } },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["capacities"],
        "summary": "Create a capacity post (supplier)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["role"],
                "properties": {
                  "role": { "type": "string" },
                  "skill_tags": { "type": "array", "items": { "type": "string" } },
                  "headcount_available": { "type": "integer", "minimum": 1, "default": 1 },
                  "availability_from": { "type": "string", "format": "date" },
                  "availability_to": { "type": "string", "format": "date" },
                  "location_city": { "type": "string" },
                  "location_postal": { "type": "string" },
                  "radius_km": { "type": "integer", "default": 50 },
                  "hourly_rate_min_cents": { "type": "integer" },
                  "hourly_rate_max_cents": { "type": "integer" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CapacityPost" } } } }
        }
      }
    },
...
  }
}